+++ Core/Models/Eft/Common/Tables/LocationServices.cs
@@ -110,16 +110,16 @@ public class BtrServerSettings
     public float? TurnCheckSensitivity { get; set; }
 
     [JsonPropertyName("DecreaseSpeedOnTurnLimit")]
-    public float? DecreaseSpeedOnTurnLimit { get; set; }
+    public double? DecreaseSpeedOnTurnLimit { get; set; }
 
     [JsonPropertyName("EndSplineDecelerationDistance")]
-    public float? EndSplineDecelerationDistance { get; set; }
+    public double? EndSplineDecelerationDistance { get; set; }
 
     [JsonPropertyName("AccelerationSpeed")]
-    public float? AccelerationSpeed { get; set; }
+    public double? AccelerationSpeed { get; set; }
 
     [JsonPropertyName("DecelerationSpeed")]
-    public float? DecelerationSpeed { get; set; }
+    public double? DecelerationSpeed { get; set; }
 
     [JsonPropertyName("PauseDurationRange")]
     public XYZ? PauseDurationRange { get; set; }
